refactor: Exception 클래스 수정
