+++ src/main/java/ddingdong/ddingdongBE/domain/form/service/FacadeCentralFormServiceImpl.java
@@ -3,8 +3,8 @@ package ddingdong.ddingdongBE.domain.form.service;
 import static ddingdong.ddingdongBE.domain.club.entity.Position.MEMBER;
 
 import ddingdong.ddingdongBE.common.exception.AuthenticationException.NonHaveAuthority;
-import ddingdong.ddingdongBE.common.exception.InvalidatedMappingException.InvalidFieldTypeException;
-import ddingdong.ddingdongBE.common.exception.InvalidatedMappingException.InvalidFormPeriodException;
+import ddingdong.ddingdongBE.common.exception.FormException.InvalidFieldTypeException;
+import ddingdong.ddingdongBE.common.exception.FormException.OverlapFormPeriodException;
 import ddingdong.ddingdongBE.domain.club.entity.Club;
 import ddingdong.ddingdongBE.domain.club.service.ClubService;
 import ddingdong.ddingdongBE.domain.clubmember.entity.ClubMember;
@@ -182,7 +182,7 @@ public class FacadeCentralFormServiceImpl implements FacadeCentralFormService {
                 endDate);
 
         if (!overlappingForms.isEmpty()) {
-            throw new InvalidFormPeriodException();
+            throw new OverlapFormPeriodException();
         }
     }
 
@@ -198,7 +198,7 @@ public class FacadeCentralFormServiceImpl implements FacadeCentralFormService {
                 .toList();
 
         if (!overlappingForms.isEmpty()) {
-            throw new InvalidFormPeriodException();
+            throw new OverlapFormPeriodException();
         }
     }
